backend/requirements.txt
fastapi==0.128.8
psycopg2-binary==2.9.10
uvicorn==0.40.0
playwright==1.58.0
APScheduler==3.11.2
python-dotenv==1.2.1
pydantic==2.12.5
curl_cffi==0.14.0
firebase-admin==7.3.0
slowapi==0.1.9
